+++ nion/utils/Selection.py
@@ -3,6 +3,7 @@ A collection of useful classes for handling selections.
 """
 
 # standard libraries
+import copy
 import enum
 import numbers
 import typing
@@ -26,58 +27,55 @@ class IndexedSelection:
         super().__init__()
         self.__changed_event = Event.Event()
         self.__indexes : typing.MutableSet[int] = set()
-        self.__anchor_index = None
+        self.__anchor_index : typing.Optional[int] = None
         self.selection_style = selection_style if selection_style else Style.multiple
 
-    def __copy__(self):
-        return type(self)(self.__indexes, self.__anchor_index)
+    def __eq__(self, other) -> bool:
+        return isinstance(other, IndexedSelection) and self.indexes == other.indexes and self.anchor_index == other.anchor_index
 
-    def __eq__(self, other):
-        return other is not None and self.indexes == other.indexes and self.anchor_index == other.anchor_index
-
-    def __ne__(self, other):
-        return other is None or self.indexes != other.indexes or self.anchor_index != other.anchor_index
+    def __ne__(self, other) -> bool:
+        return not isinstance(other, IndexedSelection) or self.indexes != other.indexes or self.anchor_index != other.anchor_index
 
     @property
-    def changed_event(self):
+    def changed_event(self) -> Event.Event:
         return self.__changed_event
 
     @property
-    def current_index(self):
+    def current_index(self) -> typing.Optional[int]:
         if len(self.__indexes) == 1:
             for index in self.__indexes:
                 return index
         return None
 
     @property
-    def anchor_index(self):
+    def anchor_index(self) -> typing.Optional[int]:
         return self.__anchor_index
 
     @anchor_index.setter
-    def anchor_index(self, value):
+    def anchor_index(self, value: typing.Optional[int]) -> None:
         self.__anchor_index = value
 
     @property
-    def has_selection(self):
+    def has_selection(self) -> bool:
         return len(self.__indexes) > 0
 
-    def contains(self, index):
+    def contains(self, index: int) -> bool:
         return index in self.__indexes
 
     @property
-    def indexes(self):
-        return self.__indexes
+    def indexes(self) -> typing.Set[int]:
+        return typing.cast(typing.Set[int], self.__indexes)
 
     @property
-    def ordered_indexes(self):
+    def ordered_indexes(self) -> typing.Sequence[int]:
         indexes = sorted(self.__indexes)
-        if len(self.__indexes) > 0 and self.__anchor_index in self.__indexes:
+        if len(self.__indexes) > 0 and self.__anchor_index is not None and self.__anchor_index in self.__indexes:
             indexes.remove(self.__anchor_index)
             indexes.insert(0, self.__anchor_index)
         return indexes
 
-    def clear(self):
-        old_index = self.__indexes.copy()
+    def clear(self) -> None:
+        old_index = copy.copy(self.__indexes)
         if self.selection_style == Style.single:
             self.__indexes.add(0)
             self.__anchor_index = 0
@@ -87,17 +85,17 @@ class IndexedSelection:
         if old_index != self.__indexes:
             self.__changed_event.fire()
 
-    def __update_anchor_index(self):
+    def __update_anchor_index(self) -> None:
         for index in self.__indexes:
             if self.__anchor_index is None or index < self.__anchor_index:
                 self.__anchor_index = index
 
-    def add(self, index):
+    def add(self, index: int) -> None:
         assert isinstance(index, numbers.Integral)
         if self.selection_style in (Style.single, Style.single_or_none):
             self.set(index)
         elif self.selection_style in (Style.multiple, ):
-            old_index = self.__indexes.copy()
+            old_index = copy.copy(self.__indexes)
             self.__indexes.add(index)
             if len(old_index) == 0:
                 self.__anchor_index = index
@@ -106,10 +104,10 @@ class IndexedSelection:
         else:
             self.clear()
 
-    def remove(self, index):
+    def remove(self, index: int) -> None:
         assert isinstance(index, numbers.Integral)
         if self.selection_style in (Style.multiple, Style.single_or_none):
-            old_index = self.__indexes.copy()
+            old_index = copy.copy(self.__indexes)
             self.__indexes.remove(index)
             if not self.__anchor_index in self.__indexes:
                 self.__update_anchor_index()
@@ -120,13 +118,13 @@ class IndexedSelection:
         else:
             self.clear()
 
-    def add_range(self, range):
-        for index in range:
+    def add_range(self, _range: range) -> None:
+        for index in _range:
             self.add(index)
 
-    def set_multiple(self, indexes):
+    def set_multiple(self, indexes: typing.Set[int]) -> None:
         if self.selection_style in (Style.multiple, ):
-            old_index = self.__indexes.copy()
+            old_index = copy.copy(self.__indexes)
             self.__indexes = set()
             self.__indexes.update(indexes)
             self.__anchor_index = list(indexes)[0] if len(indexes) > 0 else None
@@ -137,10 +135,10 @@ class IndexedSelection:
         else:
             self.clear()
 
-    def set(self, index):
+    def set(self, index: int) -> None:
         assert isinstance(index, numbers.Integral)
         if self.selection_style in (Style.multiple, Style.single, Style.single_or_none):
-            old_index = self.__indexes.copy()
+            old_index = copy.copy(self.__indexes)
             self.__indexes = set()
             self.__indexes.add(index)
             self.__anchor_index = index
@@ -149,23 +147,27 @@ class IndexedSelection:
         else:
             self.clear()
 
-    def toggle(self, index):
+    def toggle(self, index: int) -> None:
         assert isinstance(index, numbers.Integral)
         if index in self.__indexes:
             self.remove(index)
         else:
             self.add(index)
 
-    def extend(self, index):
+    def extend(self, index: int) -> None:
         assert isinstance(index, numbers.Integral)
         if self.selection_style in (Style.multiple, ):
-            old_index = self.__indexes.copy()
-            if index > self.__anchor_index:
-                for new_index in range(self.__anchor_index, index + 1):
-                    self.__indexes.add(new_index)
-            elif index < self.__anchor_index:
-                for new_index in range(index, self.__anchor_index + 1):
-                    self.__indexes.add(new_index)
+            old_index = copy.copy(self.__indexes)
+            if self.__anchor_index is not None:
+                if index > self.__anchor_index:
+                    for new_index in range(self.__anchor_index, index + 1):
+                        self.__indexes.add(new_index)
+                elif index < self.__anchor_index:
+                    for new_index in range(index, self.__anchor_index + 1):
+                        self.__indexes.add(new_index)
+            else:
+                self.__anchor_index = index
+                self.__indexes.add(index)
             if old_index != self.__indexes:
                 self.__changed_event.fire()
         elif self.selection_style in (Style.single, Style.single_or_none):
@@ -173,7 +175,7 @@ class IndexedSelection:
         else:
             self.clear()
 
-    def insert_index(self, new_index):
+    def insert_index(self, new_index: int) -> None:
         new_indexes = set()
         for index in self.__indexes:
             if index < new_index:
@@ -186,7 +188,7 @@ class IndexedSelection:
         if self.__indexes != new_indexes:
             self.__indexes = new_indexes
 
-    def remove_index(self, remove_index):
+    def remove_index(self, remove_index: int) -> None:
         new_indexes = set()
         for index in self.__indexes:
             if index != remove_index: